TESTS ON LOCAL WORKING on Travis F

+++ src/test/java/urlshortener/repository/ClickRepositoryTests.java
@@ -10,6 +10,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.test.context.junit4.SpringRunner;
 import urlshortener.domain.Click;
 import urlshortener.fixtures.ClickFixture;
@@ -29,25 +30,31 @@ public class ClickRepositoryTests {
     shortUrlRepository.save(ShortURLFixture.url2());
   }
 
+  public void cleanRepository() {
+    if (repository.count() > 0L) {
+      repository.deleteAll();
+    }
+  }
+
   @Test
   public void thatSavePersistsTheClickURL() {
-    repository.deleteAll();
+    cleanRepository();
     Click click = repository.save(ClickFixture.click(ShortURLFixture.url1()));
     assertSame(repository.count(), 1L);
     assertNotNull(click);
     assertNotNull(click.getId());
   }
 
-  @Test
+  @Test(expected = DataIntegrityViolationException.class)
   public void thatErrorsInSaveReturnsNull() {
-    repository.deleteAll();
+    cleanRepository();
     assertNull(repository.save(ClickFixture.click(ShortURLFixture.badUrl())));
-    assertSame(repository.count(), 0L);
+    assertEquals(repository.count(), 0L);
   }
 
   @Test
   public void thatFindByKeyReturnsAURL() {
-    repository.deleteAll();
+    cleanRepository();
     repository.save(ClickFixture.click(ShortURLFixture.url1()));
     repository.save(ClickFixture.click(ShortURLFixture.url2()));
     repository.save(ClickFixture.click(ShortURLFixture.url1()));
@@ -59,7 +66,7 @@ public class ClickRepositoryTests {
 
   @Test
   public void thatFindByKeyReturnsEmpty() {
-    repository.deleteAll();
+    cleanRepository();
     repository.save(ClickFixture.click(ShortURLFixture.url1()));
     repository.save(ClickFixture.click(ShortURLFixture.url2()));
     repository.save(ClickFixture.click(ShortURLFixture.url1()));
@@ -70,7 +77,7 @@ public class ClickRepositoryTests {
 
   @Test
   public void thatDeleteDelete() {
-    repository.deleteAll();
+    cleanRepository();
     Long id1 = repository.save(ClickFixture.click(ShortURLFixture.url1())).getId();
     Long id2 = repository.save(ClickFixture.click(ShortURLFixture.url2())).getId();
     repository.deleteById(id1);
